+++ src/test/java/org/opensearch/neuralsearch/stats/info/InfoStatsManagerTests.java
@@ -8,6 +8,8 @@ import org.junit.Before;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.opensearch.Version;
+import org.opensearch.neuralsearch.processor.normalization.L2ScoreNormalizationTechnique;
+import org.opensearch.neuralsearch.processor.normalization.MinMaxScoreNormalizationTechnique;
 import org.opensearch.neuralsearch.settings.NeuralSearchSettingsAccessor;
 import org.opensearch.neuralsearch.stats.common.StatSnapshot;
 import org.opensearch.neuralsearch.util.NeuralSearchClusterUtil;
@@ -18,6 +20,7 @@ import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Consumer;
 
 import static org.mockito.Mockito.when;
 
@@ -52,5 +55,28 @@ public class InfoStatsManagerTests extends OpenSearchTestCase {
         assertEquals(1, stats.size());
         assertTrue(stats.containsKey(InfoStatName.CLUSTER_VERSION));
         assertEquals("3.0.0", ((SettableInfoStatSnapshot<?>) stats.get(InfoStatName.CLUSTER_VERSION)).getValue());
+    }
+
+    public void test_callNullableIncrementer() {
+        // Create stats map with two techniques
+        Map<InfoStatName, CountableInfoStatSnapshot> stats = Map.of(
+            InfoStatName.NORM_TECHNIQUE_L2_PROCESSORS,
+            new CountableInfoStatSnapshot(InfoStatName.NORM_TECHNIQUE_L2_PROCESSORS),
+            InfoStatName.NORM_TECHNIQUE_MINMAX_PROCESSORS,
+            new CountableInfoStatSnapshot(InfoStatName.NORM_TECHNIQUE_MINMAX_PROCESSORS)
+        );
+
+        // Create incrementer map with only 1 technique
+        Map<String, Consumer<Map<InfoStatName, CountableInfoStatSnapshot>>> incrementerMap = Map.of(
+            L2ScoreNormalizationTechnique.TECHNIQUE_NAME,
+            statsParam -> infoStatsManager.increment(statsParam, InfoStatName.NORM_TECHNIQUE_L2_PROCESSORS)
+        );
+
+        // Calling nullable incrementer should only increment the technique in the map
+        infoStatsManager.callNullableIncrementer(stats, incrementerMap.get(L2ScoreNormalizationTechnique.TECHNIQUE_NAME));
+        infoStatsManager.callNullableIncrementer(stats, incrementerMap.get(MinMaxScoreNormalizationTechnique.TECHNIQUE_NAME));
+
+        assertEquals(1, (long) stats.get(InfoStatName.NORM_TECHNIQUE_L2_PROCESSORS).getValue());
+        assertEquals(0, (long) stats.get(InfoStatName.NORM_TECHNIQUE_MINMAX_PROCESSORS).getValue());
     }
 }